# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TopShelf Teaching API
  description: |
    REST API for the TopShelf Teaching educational platform.

    ## Authentication
    All protected endpoints require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Rate Limiting
    - 100 requests per minute per user
    - Rate limit headers included in all responses

    ## Errors
    All errors return a consistent JSON structure with `error`, `message`, and `code` fields.
  version: 1.0.0
  contact:
    name: TopShelf Service LLC
    email: api@topshelfservice.com
    url: https://topshelfteaching.com
  license:
    name: Proprietary
    url: https://topshelfteaching.com/license
  x-logo:
    url: https://topshelfteaching.com/logo.png
servers:
  - url: https://api.topshelfteaching.com/api/v1
    description: Production
  - url: https://staging-api.topshelfteaching.com/api/v1
    description: Staging
  - url: http://localhost:3000/api/v1
    description: Local Development
...
paths:
...
  /content/packs:
    get:
      tags: [Content]
      summary: List available content packs
      operationId: listContentPacks
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, review, approved, published, archived]
        - name: certification
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of content packs
          content:
            application/json:
              schema:
                type: object
                properties:
                  packs:
                    type: array
                    items:
                      $ref: '#/components/schemas/ContentPackSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /content/packs/{packId}:
    get:
      tags: [Content]
      summary: Get content pack details
      operationId: getContentPack
      security:
        - bearerAuth: []
      parameters:
        - name: packId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Content pack details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentPack'
        '404':
          $ref: '#/components/responses/NotFound'
  /content/packs/{packId}/blocks/{blockId}:
    get:
      tags: [Content]
      summary: Get specific block content
      operationId: getContentBlock
      security:
        - bearerAuth: []
      parameters:
        - name: packId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: blockId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Content block details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentBlock'
  /content/next/{contentPackId}:
    get:
      tags: [Content]
      summary: Get next recommended block
      operationId: getNextBlock
      security:
        - bearerAuth: []
      parameters:
        - name: contentPackId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Next block recommendation
          content:
            application/json:
              schema:
                type: object
                properties:
                  complete:
                    type: boolean
                  nextBlock:
                    $ref: '#/components/schemas/ContentBlockSummary'
                  progress:
                    type: object

  # ===========================================================================
  # BADGES
  # ===========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error, message, code]
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        requestId:
          type: string
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
...
    LearningMode:
      type: string
      enum: [L1_RECALL, L2_EXPLAIN, L3_APPLY, L4_ANALYZE, L5_EXPERT]
...
    ContentPackSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        version:
          type: string
        title:
          type: string
        description:
          type: string
        certificationTarget:
          type: string
        status:
          type: string
    ContentPack:
      allOf:
        - $ref: '#/components/schemas/ContentPackSummary'
        - type: object
          properties:
            totalBlocks:
              type: integer
            blocks:
              type: array
              items:
                $ref: '#/components/schemas/ContentBlockSummary'
    ContentBlockSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        blockId:
          type: string
        title:
          type: string
        objective:
          type: string
        targetMode:
          $ref: '#/components/schemas/LearningMode'
        sequenceOrder:
          type: integer
    ContentBlock:
      allOf:
        - $ref: '#/components/schemas/ContentBlockSummary'
        - type: object
          properties:
            timeBudgetSeconds:
              type: integer
            content:
              type: object
            hints:
              type: array
            variants:
              type: array
...